Fold the grouped-MLP validation and public wrappers into the ops module
Mirror the gated-act (SwiGLU) + MXFP8 PR's single-module structure: the
metadata-validation helpers (grouped_mlp_validation.py) and the public
wrapper module (moe_training/mxfp8_grouped_mlp.py) fold into
kernels/mxfp8/grouped_mlp_ops.py, which is now self-contained --
availability probe, is_supported, condensed validation helpers, the
four custom ops with their fakes, and the availability-gated public
wrappers. Both package __init__s return to their upstream state
(importing grouped_mlp_ops registers the ops; consumers import it
directly). All validation logic and error messages are unchanged;
docstrings condensed to contracts. Net: 3 modules / 1,363 lines ->
1 module / 1,196 lines.

29/29 tests green (rejection-message and opt-in offsets gates
included); ruff 0.11.6 lint+format clean.
