+++ torchao/prototype/moe_training/__init__.py
@@ -1,15 +1,6 @@
 from torchao.prototype.moe_training.fp8_grouped_mm import (
     _to_fp8_rowwise_then_scaled_grouped_mm,
 )
-from torchao.prototype.moe_training.mxfp8_grouped_mlp import (
-    is_supported as mxfp8_grouped_mlp_is_supported,
-)
-from torchao.prototype.moe_training.mxfp8_grouped_mlp import (
-    mxfp8_grouped_gemm,
-    mxfp8_grouped_gemm_dswiglu_bwd,
-    mxfp8_grouped_gemm_swiglu_fwd,
-    mxfp8_grouped_gemm_wgrad,
-)
 from torchao.prototype.moe_training.mxfp8_grouped_mm import (
     _to_mxfp8_then_scaled_grouped_mm,
 )
@@ -17,9 +8,4 @@ from torchao.prototype.moe_training.mxfp8_grouped_mm import (
 __all__ = [
     "_to_mxfp8_then_scaled_grouped_mm",
     "_to_fp8_rowwise_then_scaled_grouped_mm",
-    "mxfp8_grouped_gemm",
-    "mxfp8_grouped_gemm_swiglu_fwd",
-    "mxfp8_grouped_gemm_dswiglu_bwd",
-    "mxfp8_grouped_gemm_wgrad",
-    "mxfp8_grouped_mlp_is_supported",
 ]
